Basic + Safety Testing
Testing of basic usage (+ lots of fixes) and safety-checking (make sure no one
is messing with your lock while you're using it). Need to add in some concurrency
tests next, currently a placeholder for that.

Also added in some deps for testing and what not as well as some configuration changes
for running tests and some test-helpers to make things a bit easier to work with